Compute the values given the criteria and internal objects

A function that returns a named vector of evaluated criteria for each respective item in the test bank. The names are associated with the item number in the bank. Note that criteria values are returned such that the maximum value always represents the most optimal item (e.g., maximum information). In cases where the minimum value is typically selected (e.g., minimum variance) all values are multiplied by -1 to turn it into a maximization problem.

Usage

computeCriteria(
  x,
  criteria,
  person = NULL,
  test = NULL,
  design = NULL,
  subset = NULL,
  info_mats = FALSE
)

Arguments

x

an object of class 'mirtCAT_design' returned from the mirtCAT function when passing design_elements = TRUE

criteria

item selection criteria (see mirtCAT's criteria input)

person

(required when x is missing) internal person object. To be used when customNextItem function has been defined

test

(required when x is missing) internal test object. To be used when customNextItem function has been defined

design

(required when x is missing) internal design object. To be used when customNextItem function has been defined

subset

an integer vector indicating which items should be included in the optimal search; the default NULL includes all possible items. To allow only the first 10 items to be selected from this can be modified to subset = 1:10. This is useful when administering a multi-unidimensional CAT session where unidimensional blocks should be clustered together for smoother presentation. Useful when using the customNextItem function in mirtCAT

info_mats

logical; if more than one trait is present in the test, should the respective information matricies be returned instead of the scalar summary statistics (e.g., D-rule). When TRUE will return a list of matricies associated with each respective item

Value

a vector of criteria values for each respective item

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

Chalmers, R. P. (2016). Generating Adaptive and Non-Adaptive Test Interfaces for Multidimensional Item Response Theory Applications. Journal of Statistical Software, 71(5), 1-39. doi:10.18637/jss.v071.i05

See also

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

if (FALSE) { # \dontrun{
# test defined in mirtCAT help file, first example
CATdesign <- mirtCAT(df, mod, design_elements = TRUE)

computeCriteria(CATdesign, criteria = 'MI')
computeCriteria(CATdesign, criteria = 'MEI')

} # }