Function returning an object used by shiny

This function returns the GUI setup results by calling shinyApp. Primarily, this is only useful when hosting the application publicly, such as through https://www.shinyapps.io/. The function mirtCAT_preamble must be run before this function is called. The object is executed by calling runApp.

Usage

createShinyGUI(ui = NULL, host_server = TRUE)

Arguments

ui

a shiny UI function used to define the interface. If NULL, the default one will be used. See mirtCAT:::default_UI for the internal code

host_server

logical; is createShinyGUI() being used on a remote server or executed locally? When TRUE any calls to stopApp are suppressed to allow for multiple sessions to be executed. Note that FALSE gives the same behaviour as the GUI in mirtCAT

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

Chalmers, R. P. (2016). Generating Adaptive and Non-Adaptive Test Interfaces for Multidimensional Item Response Theory Applications. Journal of Statistical Software, 71(5), 1-39. doi:10.18637/jss.v071.i05

See also

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

if (FALSE) { # \dontrun{

mirtCAT_preamble(df=df)
runApp(createShinyGUI(host_server = FALSE), port = 8000) # run locally

person <- getPerson()
summary(person)

runApp(createShinyGUI(), port = 8000) # for remote server hosting

} # }