Computerized Adaptive Testing with Multidimensional Item Response Theory

Computerized Adaptive Testing with Multidimensional Item Response Theory

Details

Provides tools to generate an HTML interface for creating adaptive and non-adaptive educational and psychological tests using the shiny package. Suitable for applying unidimensional and multidimensional computerized adaptive tests using item response theory methodology and for creating simple questionnaires forms to collect response data directly in R.

Users interested in the most recent version of this package can visit https://github.com/philchalmers/mirtCAT and follow the instructions for installing the package from source (additional details about installing from Github can be found at https://github.com/philchalmers/mirt). Questions regarding the package can be sent to the mirt-package Google Group, located at https://groups.google.com/forum/#!forum/mirt-package.

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

Chalmers, R. P. (2016). Generating Adaptive and Non-Adaptive Test Interfaces for Multidimensional Item Response Theory Applications. Journal of Statistical Software, 71(5), 1-39. doi:10.18637/jss.v071.i05

Author

Phil Chalmers rphilip.chalmers@gmail.com