Preamble function called by mirtCAT

This is largely an internal function called by mirtCAT, however it is made public for better use with external web-hosting interfaces (like shinyapps.io). For more information see https://shiny.rstudio.com/articles/persistent-data-storage.html for further information about saving output remotely when using shiny.

Usage

mirtCAT_preamble(..., final_fun = NULL)

Arguments

...

arguments passed to mirtCAT

final_fun

a function called just before the shiny GUI has been terminated, primarily for saving results externally with packages such as rDrop2, RAmazonS3, googlesheets, RMySQL, personal servers, and so on when applications are hosted on the web. The function must be of the form final_fun <- function(person){...}, where person is the standard output returned from mirtCAT

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

Chalmers, R. P. (2016). Generating Adaptive and Non-Adaptive Test Interfaces for Multidimensional Item Response Theory Applications. Journal of Statistical Software, 71(5), 1-39. doi:10.18637/jss.v071.i05

See also

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

if (FALSE) { # \dontrun{

mirtCAT_preamble(df = df)

} # }